+++ regex/en-us/sensor.rx
@@ -6,3 +6,5 @@ read (out )?(the )?(?P<Entity>.*) (status|state|value|sensor)
 (how|what) is the (current )?(?P<Entity>.*) (status|state|value|sensor)
 tell me (the|our) (current )?(?P<Entity>.*) (status|state|value|sensor)
 (status|state|value|sensor) (?P<Entity>.*)
+give me the value of (?P<Entity>.*)
+what is the value of thermostat (?P<Entity>.*)
