+++ vocab/en-us/DeviceTrackerKeyword.voc
@@ -1,2 +1,3 @@
-where
+where is
 locate|location
+where
