+++ vocab/en-us/SensorStatusKeyword.voc
@@ -8,3 +8,5 @@ what is the current
 what is current
 tell me the current
 tell me current
+give me the value of
+what is the value of thermostat
